+++ server/services/torque.py
@@ -89,6 +89,7 @@ class TorqueService(Service):
     def launch(self,
                task_id,
                options,
+               gpulist,
                resource,
                docker_registry,
                docker_image,