+++ server/worker.py
@@ -1,6 +1,7 @@
 import logging
 import os
 import sys
+import six
 
 from six.moves import configparser
 
@@ -31,14 +32,31 @@ services = config.load_services(cfg.get('default', 'config_dir'))
 # remove busy state from resources
 for key in redis.keys('busy:*'):
     redis.delete(key)
+# remove reserved state from resources
+for key in redis.keys('reserved:*'):
+    redis.delete(key)
 
 # On startup, add all active tasks in the work queue.
 for task_id in task.list_active(redis):
     with redis.acquire_lock(task_id):
-        if redis.hget('task:'+task_id, 'status') == 'queue':
+        status = redis.hget('task:'+task_id, 'status')
+        if status == 'queue' or status == 'allocating' or status == 'allocated':
             task.service_queue(redis, task_id, redis.hget('task:'+task_id, 'service'))
+            task.set_status(redis, 'task:'+task_id, 'queued')
         else:
             task.work_queue(redis, task_id)
+
+# Desallocate all resources that are not anymore associated to a running task
+for service in services:
+    resources = services[service].list_resources()
+    for resource in resources:
+        keyr = 'resource:%s:%s' % (service, resource)
+        running_tasks = redis.hgetall(keyr)
+        for g, task_id in six.iteritems(running_tasks):
+            with redis.acquire_lock(task_id):
+                status = redis.hget('task:'+task_id, 'status')
+                if not(status == 'running' or status == 'terminating'):
+                    redis.hdel(keyr, g)
 
 # TODO: start multiple workers here?
 worker = Worker(redis, services,