+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### Breaking changes
 * redefine more consistent routes
+* change `task_id` as positional argument for simpler client commands
 
 ### New features
 
+* possibility to launch multi-gpu tasks
 * launch error now reports error message in task log
 * define `auto` registry - resolved by launcher checking at `default_for` in registry definition
 * add friendly name to tasks, and differentiate task type